More moved code from shape
